requirements.txt
alembic==0.9.9
aniso8601==3.0.0
blinker==1.4
certifi==2018.1.18
chardet==3.0.4
click==6.7
elastic-apm==2.1.1
Flask==0.12
Flask-Migrate==2.0.3
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.6.1
httpie==0.9.9
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
Pygments==2.2.0
python-dateutil==2.7.2
python-dotenv==0.8.2
python-editor==1.0.3
pytz==2018.4
requests==2.18.4
six==1.11.0
SQLAlchemy==1.1.5
urllib3==1.22
Werkzeug==0.14.1
